data_story/requirements.txt
dash
plotly.express
pandas
statsmodels